django_api/django_requirements.txt
Django==5.1.3
django-rest-framework==0.1.0
minio==7.2.11
pyarrow==18.0.0
markdown==3.7
django-filter==24.3
pyspark==3.5.3
setuptools
pandas